Safari-609.1.11.1/LayoutTests/media/encrypted-media/clearKey/clearKey-cenc-video-playback-mse-expected.txt


EME API is supported OK
Media source is opened OK
EVENT(encrypted)
MediaKeys is created OK
EVENT(message)
EVENT(keystatuseschange)
Session: keyId=d0d1d2d3d4d5d6d7d8d9dadbdcdddedf status=usable OK
EXPECTED (video.currentTime >= '2') OK
END OF TEST
